feat : view 클래스로 이동
