+++ packages/seal/login.html
@@ -1,44 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Parang Cafe | Login</title>
+    <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon" />
+  </head>
+  <style></style>
 
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Parang Cafe | Login</title>
-  <link rel="shortcut icon" href="./favicon.ico" type="image/x-icon">
-</head>
-<style>
-
-</style>
-
-<body>
-  <div class="app">
-    <div class="login-wrapper">
-      <div class="login-page">
-        <div class="container">
-          <form class="forms hidden">
-            <!-- <label>이메일</label> -->
-            <input type="text" id="signup-email" placeholder="이메일" />
-            <!-- <label>비밀번호</label> -->
-            <input type="password" id="signup-password" placeholder="비밀번호" />
-            <!-- <label>비밀번호 확인</label> -->
-            <input type="password" id="signup-password-confirm" placeholder="비밀번호 확인" />
-            <button data-submit="signup">회원가입</button>
-            <p class="message">계정이 있으신가요? <a href="#">로그인하기</a></p>
-          </form>
-          <form class="forms">
-            <!-- <label>이메일</label> -->
-            <input type="text" id="login-email" placeholder="이메일" />
-            <!-- <label>비밀번호</label> -->
-            <input type="password" id="login-password" placeholder="비밀번호" />
-            <button data-submit="login">로그인</button>
-            <p class="message">계정이 없으신가요? <a href="#">회원가입하기</a></p>
-          </form>
-        </div>
-      </div>
-    </div>
-  </div>
-</body>
-
+  <body>
+    <div class="app"></div>
+  </body>
 </html>
